+++ src/num/geom/mod.rs
@@ -4,11 +4,14 @@
 #![doc = crate::doc_!(modules: crate::num; geom: linear, metric, shape)]
 //
 
+#[cfg(feature = "linear")]
+#[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
 pub mod linear; // Vector*, Matrix*
+#[cfg(feature = "metric")]
+#[cfg_attr(nightly_doc, doc(cfg(feature = "metric")))]
 pub mod metric; // Position, Distance, Extent, Stride...
-
-#[cfg(feature = "geom")]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "geom")))]
+#[cfg(feature = "shape")]
+#[cfg_attr(nightly_doc, doc(cfg(feature = "shape")))]
 pub mod shape; // Point, …
 
 crate::items! { // structural access: _mods, _all
@@ -19,10 +22,11 @@ crate::items! { // structural access: _mods, _all
 
     // mod _mods {}
     mod _pub_mods { #![allow(unused)]
+        #[cfg(feature = "linear")]
         pub use super::linear::_all::*;
+        #[cfg(feature = "metric")]
         pub use super::metric::_all::*;
-
-        #[cfg(feature = "geom")]
+        #[cfg(feature = "shape")]
         pub use super::shape::_all::*;
     }
     pub(super) mod _all {